remove jib-cache dir
