+++ generators/docker-compose/templates/realm-config/jhipster-realm.json.ejs
@@ -1024,8 +1024,8 @@
     {
       "id": "6e8deddb-b4d6-4e2e-b389-b397d3f74fcd",
       "clientId": "web_app",
-      "rootUrl": "http://localhost:<%= serverPort %>",
-      "adminUrl": "http://localhost:<%= serverPort %>",
+      "rootUrl": "http://localhost:<%= entryPort %>",
+      "adminUrl": "http://localhost:<%= entryPort %>",
       "surrogateAuthRequired": false,
       "enabled": true,
       "alwaysDisplayInConsole": false,
