+++ generators/docker-prompts.js
@@ -21,7 +21,6 @@ const shelljs = require('shelljs');
 const { loadConfigs, setClusteredApps } = require('./docker-base');
 const { getBase64Secret } = require('./utils');
 const { MICROSERVICE, MONOLITH, GATEWAY } = require('../jdl/jhipster/application-types');
-const { COUCHBASE, MONGODB } = require('../jdl/jhipster/database-types');
 const { PROMETHEUS } = require('../jdl/jhipster/monitoring-types');
 const monitoring = require('../jdl/jhipster/monitoring-types');
